test(inspect): extract shared log-list e2e scenario fixture
log-list-filters and top-level-views each carried the same three-task
log directory data, MSW handler setup, and grid locators; move them to
e2e/fixtures/log-list-scenario.ts (typed with NetworkFixture instead of
the Parameters<...> gymnastics). Verified: both specs pass (28 tests).
